feat : 댓글 삭제 기능 추가

+++ src/main/java/com/example/schedule/service/JpaCommonEntityService.java
@@ -16,7 +16,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 import java.util.Map;
 
-
 @Service
 @RequiredArgsConstructor
 public class JpaCommonEntityService implements CommonEntityService {
@@ -34,18 +33,21 @@ public class JpaCommonEntityService implements CommonEntityService {
         Long userId = userRepo.save(user).getUserId();
         return Map.of("userId", userId);
     }
+
     @Transactional
     @Override
     public void authUser(UserAuthRequestDto dto) {
         User user = userRepo.findById(dto.getUserId()).orElseThrow(() -> new ResponseStatusException(HttpStatus.UNAUTHORIZED, "계정이 틀렸습니다"));
         if(!passwordEncoder.matches(dto.getPassword(),user.getPassword())) throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "비밀번호가 틀렸습니다.");
     }
+
     @Transactional
     @Override
     public void deleteUser(UserAuthRequestDto dto, Long userId) {
         authUser(dto);
         userRepo.deleteById(dto.getUserId());
     }
+
     @Transactional
     @Override
     public UserResponseDto modifyUser(UserSaveRequestDto dto, Long userId) {
@@ -57,6 +59,7 @@ public class JpaCommonEntityService implements CommonEntityService {
         //플러시를 안해주면 dto 에 수정시간이 반영되지 않음
         return modelMapper.map(user,UserResponseDto.class);
     }
+
     @Transactional
     @Override
     public UserResponseDto findUser(Long userId) {
@@ -64,6 +67,7 @@ public class JpaCommonEntityService implements CommonEntityService {
                             .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNAUTHORIZED, "계정이 존재하지 않습니다."));
         return modelMapper.map(user, UserResponseDto.class);
     }
+
     @Transactional
     @Override
     public ScheduleResponseDto createSchedule(ScheduleSaveRequestDto dto, Long userId) {
@@ -73,14 +77,18 @@ public class JpaCommonEntityService implements CommonEntityService {
         return modelMapper.map(scheduleRepo.save(schedule), ScheduleResponseDto.class)
                           .setName(user.getName());
     }
+
     @Transactional
     @Override
     public void deleteSchedule(Long userId, Long scheduleId) {
         Schedule schedule = scheduleRepo.findById(scheduleId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "일정이 존재하지 않습니다."));
+        Long check_userId = schedule.getSchedule_user().getUserId();
+        if(!userId.equals(check_userId)) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "해당 게시글의 주인이 아닙니다." );
         User user = userRepo.findById(userId).get();
         user.removeSchedule(schedule);
         scheduleRepo.deleteById(scheduleId);
     }
+
     @Transactional
     @Override
     public CommentResponseDto createComment(CommentSaveRequestDto dto, Long userId, Long scheduleId) {
@@ -94,6 +102,21 @@ public class JpaCommonEntityService implements CommonEntityService {
         return modelMapper.map(commentRepo.save(comment),CommentResponseDto.class)
                           .setUserId(user.getUserId())
                           .setName(user.getName());
+    }
+
+    @Transactional
+    @Override
+    public void deleteComment(Long userId, Long scheduleId, Long commentId) {
+        Comment comment = commentRepo.findById(commentId).orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND, "해당 댓글이 존재하지 않습니다." ));
+        Long check_userId = comment.getComment_user().getUserId();
+        if(!userId.equals(check_userId)) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "해당 댓글의 주인이 아닙니다." );
+        Schedule schedule = scheduleRepo.findById(scheduleId).orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND, "해당 일정이 존재하지 않습니다." ));
+        Long check_scheduleId = comment.getComment_schedule().getScheduleId();
+        if(!scheduleId.equals(check_scheduleId)) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "해당 게시글의 댓글이 아닙니다." );
+        User user = userRepo.findById(userId).get();
+        schedule.removeComment(comment);
+        user.removeComment(comment);
+        commentRepo.deleteById(commentId);
     }
 
 }
